Escape angle brackets
While usually Markdown should be able to distinguish between angle
brackets that are part of HTML that should be rendered directly,
and angle brackets acting as "greater than" and "less than" signs,
JSX is not as forgiving. In particular, this plays out when
rendering generics into MDX files (as used by Docusaurus 2).

Since typedoc-plugin-markdown does not attempt to render HTML
directly, we can safely convert all angle brackets to the HTML
entities (which will render as angle brackets either way.